Catch `IllegalArgumentException` thrown by `SecretKeySpec`
The `IllegalArgumentException` was thrown when the AES key passed
to `SecretKeySpec` was empty in `GiniCryptoPreAndroidM`.

This was unexpected since we always return an AES key by generating
a new one when not available. If the existing AES key could not be
decrypted then the appropriate exception was caught.

Due to a race condition the decryption `Cipher`'s  `doFinal` method
could return an empty byte array.

This happened only in the test case where we delete the RSA key pair
from the `KeyStore` and then immediately proceed with decryption.

Most of the time the expected decryption related exception was thrown,
but sometimes decryption failed silently by returning an empty
decryption result byte array.

Under normal circumstances this cannot happen in production. Clients
cannot delete the RSA key pair we use as they don't have access to
our `KeyStore` and to our secret key alias.